// pleace-golf-api/src/core/errors/error-type.enum.ts
/**
 * Copyright (c) 2020 Codev Technologies (Pty) Ltd. All rights reserved.
 */

export enum ErrorType {
    Validation = "Validation",
    ReferenceNotFound = "ReferenceNotFound",
    Range = "Range",
    Internal = "Internal",
    EmptyOrNull = "EmptyOrNull",
    Authentication = "Authentication",
    Exists = "Exists",
    DoesNotExist = "DoesNotExist",
    InvalidCode = "InvalidCode",
    Generic = "Generic",
    InviteDoesNotExist = "InviteDoesNotExist",
    InviteeAlreadyExist = "InviteeAlreadyExist"
}